+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Unreleased
 
 - LLM Scanner: Add `value_to_float` option to for converting model reported values to numeric.
+- Rename `--results` option to `--scans`. 
 
 ## 0.4.5 (03 January 2026)